This seems to allow for scanning, presuming that it is compiled without tests
It should be noted that at this point in time, extensive tests have not been performed.
Additionally, the entire test folder needs to be deleted in order to compile. The tests
will eventually need brought up to 7.6 compatibility. The Cognitive Complexity rule has
also been removed at this time. It may be reintoduced if the related issues can be resolved.